Restructure container/component file locations
