+++ src/containers/contributions/DeviceCheckBoxes.js
@@ -5,7 +5,7 @@
 import React from 'react';
 import {StyleSheet, View} from 'react-native';
 import {Colors} from '../../styles';
-import CheckBox from './CheckBox';
+import CheckBox from '../../components/CheckBox';
 
 const DeviceCheckBoxes: () => React$Node = ({
   devices,
